Move the events to another package
