+++ src/main/java/com/github/devcyntrix/deathchest/api/event/DeathChestSpawnEvent.java
@@ -1,18 +1,25 @@
-package com.github.devcyntrix.deathchest;
+package com.github.devcyntrix.deathchest.api.event;
 
 import com.github.devcyntrix.deathchest.api.DeathChest;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
 
-public class DeathChestDestroyEvent extends Event {
+public class DeathChestSpawnEvent extends Event {
 
     private static final HandlerList HANDLER_LIST = new HandlerList();
 
+    private final Player player;
     private final DeathChest deathChest;
 
-    public DeathChestDestroyEvent(DeathChest deathChest) {
+    public DeathChestSpawnEvent(Player player, DeathChest deathChest) {
+        this.player = player;
         this.deathChest = deathChest;
+    }
+
+    public Player getPlayer() {
+        return player;
     }
 
     public DeathChest getDeathChest() {
